Squashed 'src/deps/miniasync/' changes from 93283586e..dd06ea790
dd06ea790 Merge pull request #54 from lplewa/vdm
9c6390dcb masync: make vdm and future header only
e23ab21b1 Merge pull request #52 from wlemkows/KW-fixes
ad4651146 common: fix KW issues
bab51d2fc Merge pull request #51 from DamianDuy/fixNonStandardExtensionWinWarnings
0c232c340 common: remove unnecessary wrappers for allocators
d2ca2230f Merge pull request #49 from DamianDuy/fixTypeMismatchWinWarnings
a53362da8 common: fix type mismatch warnings on Windows

git-subtree-dir: src/deps/miniasync
git-subtree-split: dd06ea7907b27b8cdafaf957488e74dd1cb921a2

+++ src/core/os_thread_windows.c
@@ -565,9 +565,9 @@ void
 os_cpu_set(size_t cpu, os_cpu_set_t *set)
 {
 	internal_os_cpu_set_t *internal_set = (internal_os_cpu_set_t *)set;
-	int sum = 0;
-	int group_max = GetActiveProcessorGroupCount();
-	int group = 0;
+	size_t sum = 0;
+	WORD group_max = GetActiveProcessorGroupCount();
+	WORD group = 0;
 	while (group < group_max) {
 		sum += GetActiveProcessorCount(group);
 		if (sum > cpu) {
